[semantic-arc] Split out copy_value optimizations (except phi elimination) into its own file.
