+++ lib/SILOptimizer/SemanticARC/CMakeLists.txt
@@ -2,4 +2,5 @@ target_sources(swiftSILOptimizer PRIVATE
   SemanticARCOpts.cpp
   OwnershipLiveRange.cpp
   LoadCopyToLoadBorrowOpt.cpp
-  BorrowScopeOpts.cpp)
+  BorrowScopeOpts.cpp
+  CopyValueOpts.cpp)
